Added comments to code and a switch scene function

+++ src/main/java/com/example/nea/Game.java
@@ -47,20 +47,26 @@ public class Game {
             @Override
             public void handle(long l) {
 
+                // calculate camera x and y coordinates relative to player
                 cam_x = (int) (player.map_x - ((double) main.width /2));
                 cam_y = (int) (player.map_y - ((double) main.height /2));
 
+                // calculate maximum camera positions for bound checking
                 int max_cam_x = (bg.tilemap[0].length*bg.tile_width) - main.width;
                 int max_cam_y = (bg.tilemap.length*bg.tile_width) - main.height;
 
+                // adjust camera in case it forces the displayed background to go beyond the given walls.
                 cam_x = checkCam(cam_x, max_cam_x);
                 cam_y = checkCam(cam_y, max_cam_y);
 
+                // update the player's position
                 player.updatePlayer(); // handle player movement and collision
 
-                bg.renderCanvas((int) (cam_y/ bg.tile_width), (int) (cam_x/ bg.tile_width));
+                // render the canvas based of the camera position
+                bg.renderCanvas((cam_y/ bg.tile_width), (int) (cam_x/ bg.tile_width));
             }
 
+            // function to check camera angles
             private int checkCam(int cam, int max) {
                 if (cam < 0){
                     return 0;
@@ -68,7 +74,6 @@ public class Game {
                     return max;
                 }
                 return cam;
-
             }
         };
         timer.start();
